requirements.txt
transformers
torch
scikit-learn
tensorflow
fastapi
uvicorn[standard]
python-multipart
pillow
Jinja2
SQLAlchemy
psycopg2-binary
websockets
python-socketio
opencv-python
moviepy
speechrecognition
pydub
requests
beautifulsoup4
selenium
fake-useragent
